requirement.txt
streamlit
PyPDF2
faiss-cpu
langchain
langchain-community
langchain-groq
huggingface-hub
sentence-transformers
python-dotenv
gTTS
moviepy
